% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFDR.R
\name{callFDR}
\alias{callFDR}
\title{callFDR}
\usage{
callFDR(foreground, background, thresList = NULL, adjust = T, verbose = F)
}
\arguments{
\item{foreground}{a vector of numeric values, either IPD ratio or Score from the native DNA}

\item{background}{a vector of numeric values, either IPD ratio or Score from the WGA}

\item{thresList}{a vector of numeric values indicating the threshold list for FDR calling}

\item{adjust}{a boolean value indicating whether the estimated FDRs are adjusted to be decreased with the increased thresholds}

\item{verbose}{a boolean value indicating whether output on screen}
}
\value{
a data frame values comprsing two columns
\itemize{
 \item {thres} a vector of numeric values indicating the threshold list 
 \item {fdr} a vector of numeric values indicating the estimated FDRs correpsonding to the thresholds
}
}
\description{
callFDR calls False Discovery Rate by comparing the foreground values with the background values
}
\examples{


}
