% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMotifHeatmap.R
\name{drawMotifHeatmap}
\alias{drawMotifHeatmap}
\title{drawMotifHeatmap}
\usage{
drawMotifHeatmap(motif_fdr, outputName)
}
\arguments{
\item{motif_fdr}{a data frame comprising the follow components
\itemize{
 \item {motif} a character value indicating the methylation motifs
 \item {pos} an integer indicating the methylation position in the motif
 \item {score} a numeric value indicating the motif enrichment score for the corresponding motif
}}

\item{outputName}{a character value indicating the output pdf file name}
}
\description{
drawMotifHeatmap plots the heatmap of motif enrichment scores
}
\examples{

}
