% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerateMotif.R
\name{enumerateMotif}
\alias{enumerateMotif}
\title{enumerateMotif}
\usage{
enumerateMotif(kmer, meth_char)
}
\arguments{
\item{kmer}{an integer value indicating the motif length}

\item{meth_char}{a character value indicating the methylated position in the motif}
}
\value{
a data frame comprising the following components
\itemize{
 \item {motif} a character value indicating the methylation motifs
 \item {pos} an integer indicating the methylation position in the motif
}
}
\description{
enumerateMotif enumerates all k-mer motifs
}
\examples{

}
