% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSmrtOfMotif.R
\name{getSmrtOfMotif}
\alias{getSmrtOfMotif}
\title{getSmrtOfMotif}
\usage{
getSmrtOfMotif(smrt, motif, methy_pos, genome)
}
\arguments{
\item{smrt}{a data frame of smrt data comprising at least refName, strand and tpl}

\item{motif}{a character value indicating the methylation motif}

\item{methy_pos}{an integer value indicating the methylated position in the methylation motif}

\item{genome}{a DNAStringSet of the genome reference}
}
\value{
a data frame of smrt data comprising of the methylation positions
}
\description{
getSmrtOfMotif extracts the rows in a smrt data for the methylation position in a given methylation motif
}
\examples{

}
