% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemiMethylation.R
\name{hemiMethylation}
\alias{hemiMethylation}
\title{hemiMethylation}
\usage{
hemiMethylation(SMipdRatio, draw = FALSE, point = c("plot",
  "smoothScatter"))
}
\arguments{
\item{SMipdRatio}{a list with the name of refName and strand "refName.strand". Each item is a data frame comprising the following columns:
\itemize{
 \item {tpl} an integer value indicating the genomic location 
 \item {ipdRatio} a numeric value indicating the IPD ratio at this position of one specific molecule
}}

\item{draw}{a boolean value indicating whehter draw the figure or not}
}
\value{
a data frame comprising the following columns:
\itemize{
 \item {pos_ipdRatio} a vector of numeric values indicating the IPD ratio on the positive strand
 \item {neg_ipdRatio} a vector of numeric values indicating the IPD ratio on the negative strand
}
}
\description{
hemiMethylation checks strand-specific methylation from the Single-molecule level
}
\examples{

}
\seealso{
\code{\link{smrterSyntheticControl}}; \code{\link{smrterTraditionalControl}};
}
