% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCmpH5.R
\name{loadCmpH5}
\alias{loadCmpH5}
\title{loadCmpH5}
\usage{
loadCmpH5(path)
}
\arguments{
\item{path}{a character value indicating the file path of cmp.h5 file}
}
\value{
a list comprising the following components
\itemize{
 \item {cmpH5} a h5 data saving the IPDs and sequence for each subread
 \item {index} a data frame saving the basic information of each subread, e.g., chromosome, start/end location in genome, and start/end location in the read 
 \item {accuracy} an vector of read accuracy of one subread
 \item {readlength} a vector of integer values indicating the subread length
}
}
\description{
loadCmpH5 loads the cmp.h5 format data
}
\examples{

}
