% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSmrtGff.R
\name{readSmrtGff}
\alias{readSmrtGff}
\title{readSmrtGff}
\usage{
readSmrtGff(file)
}
\arguments{
\item{file}{a character value indicating the file name of the gff format smrt data}
}
\value{
a data frame of smrt data comprising the following columns
\itemize{
 \item {refName} a character value indicating the chromosomes
 \item {strand} a character value indicating the strand taking on either 0 (reference) or 1 (opposite)
 \item {tpl} an integer value indicating the genomic location 
 \item {base}	a character value indiating the cognate base at this position in the reference
 \item {score} an integer value indicating the Phred-transformed pvalue of modification detection, i.e. 10*(-log10(pvalue))
 \item {coverage} an integer value indicating the count of valid IPDs for the in-silico control mode or the mean of case and control coverage for the case-control mode
 \item {ipdRatio} a numeric value indicating the ipd Ratio
 \item {tag} a character value indicating the estimated modification types, including "modified_base","m6A","m5C" and "m4C"
 \item {identificationQv} an integer value indicating the score matching the modification signitures of "m6A","m5C" and "m4C"
 \item {context}	a character value indiating the bases in [-20,+20] of this position in the reference
}
}
\description{
readSmrtGff reads the gff format smrt data (modifications.gff)
}
\examples{

}
\seealso{
\code{\link{readSmrtCsv}};
}
