% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSmrt.R
\name{splitSmrt}
\alias{splitSmrt}
\title{splitSmrt}
\usage{
splitSmrt(smrt, sorted = TRUE)
}
\arguments{
\item{smrt}{a data frame of smrt data comprising at least refName, strand, tpl}

\item{sorted}{a boolean value indicating whether the output is sorted based on tpl}
}
\value{
a list comprising of all refNames on strand of either 0 or 1. The name of each item in the list is "refName.strand".
}
\description{
splitSmrt splits the smrt data into chunks based on both refName and strand
}
\examples{


}
